requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10
SpeechRecognition>=3.10.0
pocketsphinx>=5.0.0
vosk>=0.3.45
ollama>=0.1.0
PyAudio>=0.2.13
python-xlib>=0.33
pycairo>=1.23.0
pygobject>=3.44.0
pynput>=1.7.6
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.3
aiohttp>=3.8.5
asyncio>=3.4.3 